# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: VoiceAsset Server API
  version: 0.22.0
  summary: Public REST contract for the VoiceAsset modular monolith.
  description: >-
    VoiceAsset provides local web sessions, immutable audio ingestion, durable
    transcription jobs, encrypted workspace-scoped ASR provider profiles,
    transcript retrieval, authenticated byte-range audio delivery, and bounded
    asynchronous waveform derivatives. Resource access is always scoped to the
    authenticated workspace.
  license:
    name: GNU Affero General Public License v3.0 or later
    identifier: AGPL-3.0-or-later
servers:
  - url: /
    description: Server hosting this contract
...
security: []
paths:
...
  /api/v1/hotword-sets:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
    get:
      tags: [Hotwords]
      summary: List workspace hotword sets
      description: Requires `admin:read`; each item contains its current immutable entry version.
      operationId: listHotwordSets
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/HotwordSetListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '405':
          $ref: '#/components/responses/MethodNotAllowedResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailableResponse'
    post:
      tags: [Hotwords]
      summary: Create a versioned hotword set
      description: >-
        Requires `admin:write`. Terms use a provider-neutral 1-100 weight and
        are compiled by the selected ASR adapter at job execution time.
      operationId: createHotwordSet
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHotwordSetRequest'
      responses:
        '201':
          $ref: '#/components/responses/HotwordSetCreatedResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '405':
          $ref: '#/components/responses/MethodNotAllowedResponse'
        '409':
          $ref: '#/components/responses/ConflictResponse'
        '413':
          $ref: '#/components/responses/RequestTooLargeResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailableResponse'
...
  /api/v1/glossary-sets:
    parameters:
      - $ref: '#/components/parameters/XRequestID'
    get:
      tags: [Glossaries]
      summary: List workspace correction glossary sets
      operationId: listGlossarySets
      security:
        - SessionCookie: []
        - BearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/GlossarySetListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '405':
          $ref: '#/components/responses/MethodNotAllowedResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailableResponse'
    post:
      tags: [Glossaries]
      summary: Create a versioned correction glossary
      operationId: createGlossarySet
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGlossarySetRequest'
      responses:
        '201':
          $ref: '#/components/responses/GlossarySetCreatedResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '405':
          $ref: '#/components/responses/MethodNotAllowedResponse'
        '409':
          $ref: '#/components/responses/ConflictResponse'
        '413':
          $ref: '#/components/responses/RequestTooLargeResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailableResponse'
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: voiceasset_session
      description: >-
        Opaque HttpOnly session cookie. Unsafe cookie-authenticated operations
        also require Origin to equal the configured public origin.
    RefreshCookie:
      type: apiKey
      in: cookie
      name: voiceasset_refresh
      description: >-
        Opaque HttpOnly refresh cookie scoped to `/api/v1/auth`. It is accepted
        only by the rotation endpoint together with the exact configured Origin.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: VoiceAsset opaque token
      description: >-
        Opaque session token beginning with `va_`, or scoped API key beginning
        with `va_pat_`. If both cookie and bearer credentials are sent, they
        must contain the same token. API keys cannot revoke web sessions.
  parameters:
    XRequestID:
      name: X-Request-ID
      in: header
      required: false
      description: Optional caller correlation ID; oversized values are replaced.
      schema:
        type: string
        minLength: 1
        maxLength: 200
    Origin:
      name: Origin
      in: header
      required: false
      description: >-
        Browser origin. Required and exact for unsafe cookie-authenticated
        operations; optional for bearer authentication.
      schema:
        type: string
        format: uri
...
  headers:
    XAPIVersion:
      description: Stable API namespace served by this process.
      schema:
        type: string
        const: v1
    XRequestID:
      description: Caller-provided or server-generated request correlation ID.
      schema:
        type: string
        minLength: 1
        maxLength: 200
    XServerVersion:
      description: Semantic server build version.
      schema:
        type: string
        minLength: 1
    Location:
      description: Relative URL of the created or completed resource.
      schema:
        type: string
    ETag:
      description: Quoted resource version or lowercase SHA-256 entity tag.
      schema:
        type: string
...
    WWWAuthenticate:
      description: Bearer authentication challenge.
      schema:
        type: string
        const: 'Bearer realm="voiceasset"'
...
  responses:
...
    BadRequestResponse:
      description: Invalid JSON, identifier, idempotency key, or operation input.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedResponse:
      description: Session credential is missing, malformed, expired, revoked, or ambiguous.
      headers:
        WWW-Authenticate:
          $ref: '#/components/headers/WWWAuthenticate'
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ForbiddenResponse:
      description: Origin validation failed or the authenticated principal lacks the required scope.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundResponse:
      description: Resource does not exist in the authenticated workspace.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    MethodNotAllowedResponse:
      description: The route exists but does not support this HTTP method.
      headers:
        Allow:
          description: Comma-separated methods accepted by the route.
          schema:
            type: string
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictResponse:
      description: Current resource, membership, idempotency input, upload state, or asset state conflicts with this request.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RequestTooLargeResponse:
      description: JSON or binary request body exceeds the operation limit.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalErrorResponse:
      description: Request could not be completed due to an internal failure.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailableResponse:
      description: A required application service is not configured or not ready.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    HotwordSetCreatedResponse:
      description: Version-one hotword set created.
      headers:
        Location:
          $ref: '#/components/headers/Location'
        ETag:
          $ref: '#/components/headers/ETag'
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HotwordSet'
...
    HotwordSetListResponse:
      description: Workspace hotword sets with their current immutable versions.
      headers:
        X-API-Version:
          $ref: '#/components/headers/XAPIVersion'
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
        X-Server-Version:
          $ref: '#/components/headers/XServerVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HotwordSetList'
...
    GlossarySetCreatedResponse:
      description: Version-one correction glossary created.
      headers:
        Location:
          $ref: '#/components/headers/Location'
        ETag:
          $ref: '#/components/headers/ETag'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GlossarySet'
...
    GlossarySetListResponse:
      description: Workspace correction glossaries.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GlossarySetList'
...
  schemas:
...
    CreateHotwordSetRequest:
      type: object
      additionalProperties: false
      required: [display_name, scope_type, entries]
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 100
        scope_type:
          $ref: '#/components/schemas/HotwordScopeType'
        scope_id:
          type: string
          format: uuid
          description: Required for collection and asset scopes; omitted for workspace scope.
        state:
          $ref: '#/components/schemas/HotwordState'
        entries:
          type: array
          minItems: 1
          maxItems: 500
          items:
            $ref: '#/components/schemas/HotwordEntryInput'
...
    HotwordEntryInput:
      type: object
      additionalProperties: false
      required: [term, language, weight]
      properties:
        term:
          $ref: '#/components/schemas/HotwordTerm'
        aliases:
          type: array
          maxItems: 20
          uniqueItems: true
          items:
            $ref: '#/components/schemas/HotwordTerm'
        language:
          type: string
          pattern: '^[A-Za-z]{2,8}(?:-[A-Za-z0-9]{1,8})*$'
        weight:
          type: integer
          minimum: 1
          maximum: 100
          description: Provider-neutral semantic weight compiled by each adapter.
        provider_mapping:
          $ref: '#/components/schemas/HotwordProviderMapping'
        enabled:
          type: boolean
          default: true
        description:
          type: string
          maxLength: 500
    HotwordEntry:
      type: object
      additionalProperties: false
      required: [term, aliases, language, weight, provider_mapping, enabled]
      properties:
        term:
          $ref: '#/components/schemas/HotwordTerm'
        aliases:
          type: array
          maxItems: 20
          uniqueItems: true
          items:
            $ref: '#/components/schemas/HotwordTerm'
        language:
          type: string
          pattern: '^[A-Za-z]{2,8}(?:-[A-Za-z0-9]{1,8})*$'
        weight:
          type: integer
          minimum: 1
          maximum: 100
        provider_mapping:
          $ref: '#/components/schemas/HotwordProviderMapping'
        enabled:
          type: boolean
        description:
          type: string
          maxLength: 500
    HotwordTerm:
      type: string
      minLength: 1
      maxLength: 30
      description: UTF-8 term without control characters or comma and pipe delimiters.
    HotwordProviderMapping:
      type: object
      additionalProperties: false
      maxProperties: 3
      properties:
        mock_asr:
          type: object
          maxProperties: 32
          additionalProperties: true
        aliyun_asr:
          type: object
          maxProperties: 32
          additionalProperties: true
        tencent_asr:
          type: object
          maxProperties: 32
          additionalProperties: true
    HotwordScopeType:
      type: string
      enum: [workspace, collection, asset]
    HotwordState:
      type: string
      enum: [enabled, disabled]
      default: enabled
    HotwordSet:
      type: object
      additionalProperties: false
      required:
        - id
        - workspace_id
        - display_name
        - scope_type
        - scope_id
        - state
        - current_version
        - resource_version
        - entries
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        display_name:
          type: string
          minLength: 1
          maxLength: 100
        scope_type:
          $ref: '#/components/schemas/HotwordScopeType'
        scope_id:
          type: [string, 'null']
          format: uuid
        state:
          $ref: '#/components/schemas/HotwordState'
        current_version:
          type: integer
          minimum: 1
        resource_version:
          type: integer
          format: int64
          minimum: 1
        entries:
          type: array
          minItems: 1
          maxItems: 500
          items:
            $ref: '#/components/schemas/HotwordEntry'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    HotwordSetList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 1000
          items:
            $ref: '#/components/schemas/HotwordSet'
    CreateGlossarySetRequest:
      type: object
      additionalProperties: false
      required: [display_name, scope_type, entries]
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 100
        scope_type:
          $ref: '#/components/schemas/HotwordScopeType'
        scope_id:
          type: string
          format: uuid
        state:
          type: string
          enum: [enabled, disabled]
          default: enabled
        entries:
          type: array
          minItems: 1
          maxItems: 500
          items:
            $ref: '#/components/schemas/GlossaryEntry'
...
    GlossaryEntry:
      type: object
      additionalProperties: false
      required: [canonical_form, aliases, language, priority]
      properties:
        canonical_form:
          type: string
          minLength: 1
          maxLength: 200
        aliases:
          type: array
          minItems: 1
          maxItems: 50
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 200
        language:
          type: string
          pattern: '^(?:\*|[A-Za-z]{2,8}(?:-[A-Za-z0-9]{1,8})*)$'
        context_terms:
          type: array
          maxItems: 50
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 200
        forbidden_contexts:
          type: array
          maxItems: 50
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 200
        regex:
          type: boolean
          default: false
        case_sensitive:
          type: boolean
          default: false
        priority:
          type: integer
          minimum: 1
          maximum: 1000
        description:
          type: string
          maxLength: 500
    GlossarySet:
      type: object
      additionalProperties: false
      required:
        - id
        - workspace_id
        - display_name
        - scope_type
        - scope_id
        - state
        - current_version
        - resource_version
        - entries
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        display_name:
          type: string
        scope_type:
          $ref: '#/components/schemas/HotwordScopeType'
        scope_id:
          type: [string, 'null']
          format: uuid
        state:
          type: string
          enum: [enabled, disabled]
        current_version:
          type: integer
          minimum: 1
        resource_version:
          type: integer
          format: int64
          minimum: 1
        entries:
          type: array
          minItems: 1
          maxItems: 500
          items:
            $ref: '#/components/schemas/GlossaryEntry'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    GlossarySetList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GlossarySet'
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/APIError'
    APIError:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          minLength: 1
        request_id:
          type: string
          minLength: 1
          maxLength: 200
    ErrorCode:
      type: string
      enum:
        - asset_not_ready
        - checksum_mismatch
        - correction_active
        - audit_unavailable
        - email_conflict
        - forbidden
        - glossary_conflict
        - hotword_conflict
        - idempotency_conflict
        - internal_error
        - invalid_credentials
        - invalid_pairing
        - invalid_request
        - last_owner_required
        - method_not_allowed
        - not_ready
        - not_found
        - origin_not_allowed
        - precondition_required
        - profile_conflict
        - rate_limited
        - request_too_large
        - review_conflict
        - revision_not_correctable
        - service_unavailable
        - encryption_unavailable
        - unauthorized
        - unsupported_media
        - unsupported_media_type
        - version_conflict
        - upload_conflict
        - upload_expired
        - upload_incomplete